test(prompter): skip vt10x tests on Windows
